TJFE-52 [rename] Screen.kt 패키지 이동 (utils -> common/screen)
